Melhorar o script do ambiente

+++ run.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 # Frescura
 Black='\033[0;30m'
 Red='\033[0;31m'
@@ -9,16 +11,36 @@ Cyan='\033[0;36m'
 LightGray='\033[0;37m'
 Clear='\033[0m'
 
-project=$(basename "$PWD" | tr '[:upper:]' '[:lower:]')
+usage() {
+	echo "$0 [--host] [--help]"
+	echo -e "\t--host: roda o ambiente LAMP sem o PHPMyAdmin nem a docker bridge"
+	echo -e "\t--help: mostra esse texto"
+}
+
+if [[ "$@" == *"--help"* ]]; then
+	usage
+	exit 0
+fi
+
+host_mode=false
+if [[ "$@" = *"--host"* ]]; then
+	host_mode=true
+fi
+
+project=$(basename "$PWD" | tr '[:upper:]' '[:lower:]' | tr ' ' '_')
 
 # Salvar o banco
 docker volume create database
 
-docker run -p 3306:3306 -p 443:443 --name "$project"_lamp --rm --mount 'source=database,target=/var/lib/mysql' -v "$PWD":/srv/http -e MYSQL_USER=$project -e MYSQL_PASSWORD=password -d aufgfua/lamp:run
+if [[ $host == false ]]; then
+	docker run -p 3306:3306 -p 443:443 --name "$project"_lamp --rm --mount 'source=database,target=/var/lib/mysql' -v "$PWD":/srv/http -e MYSQL_USER="$project" -e MYSQL_PASSWORD=password -d aufgfua/lamp:run
+else
+	docker run --net=host --name "$project"_lamp --rm --mount 'source=database,target=/var/lib/mysql' -v "$PWD":/srv/http -e MYSQL_USER="$project" -e MYSQL_PASSWORD=password -d aufgfua/lamp:run
+fi
 
-
-
-docker run --name "$project"_myadmin --rm --link "$project"_lamp:db -p 8000:80 -d phpmyadmin/phpmyadmin
+if [[ $host == false ]]; then
+	docker run --name "$project"_myadmin --rm --link "$project"_lamp:db -p 8000:80 -d phpmyadmin/phpmyadmin
+fi
 
 echo -e "The project ${Red}$project${Clear} is beign run..."
 echo -e "MySQL Username: ${Green}$project${Clear}"
@@ -28,4 +50,8 @@ echo -e "Press ${Purple}return${Clear} to stop docker..."
 
 read nothing
 
-docker kill lojinha_lamp lojinha_myadmin
+docker kill lojinha_lamp
+
+if [[ $host == false ]]; then
+	docker kill lojinha_myadmin
+fi
